feat: add unit factory
