+++ src/modules/unit/unit.module.ts
@@ -4,10 +4,12 @@ import { UnitController } from './unit.controller';
 import { PrismaModule } from '@/prisma/prisma.module';
 import { OrderRepository } from '@/modules/order/order.repository';
 import { TransactionRepository } from '@/modules/transaction/transaction.repository';
+import { UnitFactory } from './unit.factory';
 
 @Module({
   imports: [PrismaModule],
   controllers: [UnitController],
-  providers: [UnitService, OrderRepository, TransactionRepository],
+  providers: [UnitService, OrderRepository, TransactionRepository, UnitFactory],
+  exports: [UnitFactory],
 })
 export class UnitModule {}
